refactor(memory): extract maintenance state machine
