+++ core/memory/runtime.py
@@ -11,7 +11,8 @@ import stat
 from concurrent.futures import CancelledError as FutureCancelledError
 from concurrent.futures import Future as ThreadFuture
 from concurrent.futures import TimeoutError as FutureTimeoutError
-from collections.abc import Awaitable, Callable, Mapping
+from collections.abc import AsyncIterator, Awaitable, Callable, Mapping
+from contextlib import asynccontextmanager
 from dataclasses import asdict, replace
 from datetime import datetime, timezone
 from pathlib import Path
@@ -27,17 +28,8 @@ from core.memory.artifact import (
     MemoryRuntimeActivationError,
     get_memory_artifact_manager,
 )
-from core.memory.backup_restore_journal import (
-    BackupRestoreConflict,
-    BackupRestoreOperation,
-    MemoryBackupRestoreJournal,
-)
 from core.memory.blocking import run_blocking
-from core.memory.clear_journal import (
-    ClearOperation,
-    ClearSurface,
-    MemoryClearJournal,
-)
+from core.memory.clear_journal import ClearSurface
 from core.memory.everos import (
     EverOSPort,
     MemoryProviderFailure,
@@ -46,6 +38,14 @@ from core.memory.everos import (
 from core.memory.everos_insight import MemoryInsightPaths, MemoryInsightReader
 from core.memory.everos_insight.recorder import clear_call_log, maintain_call_log
 from core.memory.module import MemoryModule
+from core.memory.maintenance import (
+    ClearRecoveryResult,
+    ClearResult,
+    MaintenanceRuntimeState,
+    MemoryMaintenance,
+    MemoryMaintenanceRuntimePort,
+    MemoryStoreUnavailableError,
+)
 from core.memory.process import (
     EverOSProcess,
     EverOSProcessFactory,
@@ -60,7 +60,7 @@ from core.memory.provider_root import (
     ProviderRootRollback,
 )
 from core.memory.store import MemoryStore, is_principal_id, is_project_id
-from core.memory.snapshot import MemorySnapshot, MemorySnapshotManager
+from core.memory.snapshot import MemorySnapshot
 from core.memory.types import (
     MemoryItems,
     MemoryResult,
@@ -76,7 +76,6 @@ from core.memory.worker import ProcessingEvent
 logger = logging.getLogger(__name__)
 
 
-_MaintenanceIOResult = TypeVar("_MaintenanceIOResult")
 _SessionLifecycleResult = TypeVar("_SessionLifecycleResult")
 
 
@@ -90,8 +89,33 @@ def _new_recorder_episode_id() -> str:
     return f"ma_{secrets.token_hex(32)}"
 
 
-class MemoryStoreUnavailableError(RuntimeError):
-    """Raised when the controller cannot safely open the local Memory store."""
+def _clear_recovery_payload(
+    recovery: ClearRecoveryResult | None,
+) -> dict[str, Any] | None:
+    if recovery is None:
+        return None
+    return {
+        "state": recovery.state,
+        "operation_id": recovery.operation_id,
+        "occurred_at": recovery.occurred_at,
+        "error_code": recovery.error_code,
+        "can_resume": recovery.can_resume,
+        "can_abort": recovery.can_abort,
+    }
+
+
+def _clear_result_payload(result: ClearResult) -> dict[str, Any]:
+    if result.status == "failed":
+        return {
+            "status": "failed",
+            "error": result.error,
+            "recovery": _clear_recovery_payload(result.recovery),
+        }
+    return {
+        "status": result.status,
+        "operation_id": result.operation_id,
+        "epoch": result.epoch,
+    }
 
 
 class MemorySessionLifecycleBusyError(RuntimeError):
@@ -156,29 +180,7 @@ class MemoryRuntime:
         self._config = config
         self._restart_config = deepcopy(config)
         self._effective_home = effective_home or paths.get_vibe_remote_dir()
-        self._backup_active = False
-        self._backup_restore_journal: MemoryBackupRestoreJournal | None = None
-        self._clear_journal: MemoryClearJournal | None = None
-        self._snapshot_manager: MemorySnapshotManager | None = None
-        self._backup_manager: MemorySnapshotManager | None = None
-        self._clear_journal_error: Exception | None = None
-        try:
-            self._backup_restore_journal = MemoryBackupRestoreJournal(
-                self._effective_home
-            )
-            self._clear_journal = MemoryClearJournal(self._effective_home)
-            self._snapshot_manager = MemorySnapshotManager(self._effective_home)
-            self._backup_manager = MemorySnapshotManager._for_backup(
-                self._effective_home,
-                operation_guard=self._clear_journal.assert_backup_allowed,
-            )
-            self._backup_restore_journal.mark_boot_recovery_needed()
-            self._clear_journal.mark_boot_recovery_needed()
-        except Exception as exc:
-            self._clear_journal_error = exc
-            logger.exception(
-                "Memory maintenance journal initialization failed; maintenance is fenced"
-            )
+        self._maintenance: MemoryMaintenance | None = None
         self._artifact_manager: MemoryArtifactPort = artifact_manager or get_memory_artifact_manager()
         self._provider_root_owner = ProviderRoot(
             self._provider_root,
@@ -194,8 +196,6 @@ class MemoryRuntime:
         self._reconcile_lock = asyncio.Lock()
         self._restart_task: asyncio.Task[dict[str, Any]] | None = None
         self._ready_activation_task: asyncio.Task[None] | None = None
-        self._terminal_snapshot_gc_task: asyncio.Task[None] | None = None
-        self._backup_stage_reconcile_task: asyncio.Task[None] | None = None
         self._closing = False
         self._ready_event: EverOSProcessPort | None = None
         # Single-flight state for the drain-side processing gate. Deliberately a
@@ -218,6 +218,11 @@ class MemoryRuntime:
         self._recorder_health_episode_id: str | None = None
         self._last_health_snapshot: ProviderHealthSnapshot | None = None
         self._last_health_observed_at: str | None = None
+        self._maintenance = MemoryMaintenance(
+            store,
+            effective_home=self._effective_home,
+            runtime=self._maintenance_runtime_port(),
+        )
         self._open_store(store)
 
     def _open_store(self, store: MemoryStore | None = None) -> bool:
@@ -252,6 +257,7 @@ class MemoryRuntime:
             return False
         self._store = opened
         self._module = module
+        self._require_maintenance().attach_store(opened)
         if self._maintenance_open():
             module._worker.pause_claims()
         self._store_error = None
@@ -295,56 +301,62 @@ class MemoryRuntime:
     def _maintenance_open(self) -> bool:
         """Fail closed when the independent clear authority cannot prove terminal."""
 
-        if self._backup_active:
-            return True
-        restore_journal = self._backup_restore_journal
-        if restore_journal is None:
-            return True
-        try:
-            if restore_journal.get_open_operation() is not None:
-                return True
-        except Exception:
-            return True
-        journal = self._clear_journal
-        if journal is None or self._clear_journal_error is not None:
-            return True
-        try:
-            return journal.get_open_operation() is not None
-        except Exception:
-            return True
+        maintenance = self._maintenance
+        return maintenance is None or maintenance.is_open()
 
     def _can_disable_without_maintenance_authority(self, config: MemoryConfig) -> bool:
         """Allow only a pure disable when the journal authority is unreadable."""
 
         if config.enabled or config != replace(self._config, enabled=False):
             return False
-        restore_journal = self._backup_restore_journal
-        clear_journal = self._clear_journal
-        if (
-            restore_journal is None
-            or clear_journal is None
-            or self._clear_journal_error is not None
-        ):
-            return True
-        try:
-            restore_journal.get_open_operation()
-            clear_journal.get_open_operation()
-        except Exception:
-            return True
-        return False
+        maintenance = self._maintenance
+        return maintenance is None or maintenance.can_disable_without_authority()
 
-    def _clear_capability_ready(self) -> bool:
-        """Whether every local dependency required by Clear initialized."""
-
-        return (
-            self._store is not None
-            and self._module is not None
-            and self._backup_restore_journal is not None
-            and self._clear_journal is not None
-            and self._snapshot_manager is not None
-            and self._backup_manager is not None
-            and self._clear_journal_error is None
+    def _maintenance_runtime_port(self) -> MemoryMaintenanceRuntimePort:
+        return MemoryMaintenanceRuntimePort(
+            exclusive_fence=self._maintenance_fence,
+            boot_recovery_fence=self._maintenance_boot_recovery_fence,
+            state=self._maintenance_runtime_state,
+            enter_maintenance=self._enter_maintenance,
+            leave_maintenance=self._leave_maintenance,
+            pause_claims=self._pause_clear_claims,
+            resume_claims=self._resume_maintenance_claims,
+            quiesce=self._quiesce_for_clear,
+            resume=self._resume_after_clear,
+            delete_surface=self._delete_clear_surface,
         )
+
+    @asynccontextmanager
+    async def _maintenance_fence(self) -> AsyncIterator[None]:
+        """Acquire destructive fences in their single permitted order."""
+
+        async with self._reconcile_lock, self.module._lifecycle_lock:
+            async with self.module._root_lifecycle_lock():
+                yield
+
+    @asynccontextmanager
+    async def _maintenance_boot_recovery_fence(self) -> AsyncIterator[None]:
+        """Add the root fence while reconcile and module fences are already held."""
+
+        async with self.module._root_lifecycle_lock():
+            yield
+
+    def _maintenance_runtime_state(self) -> MaintenanceRuntimeState:
+        return MaintenanceRuntimeState(
+            artifact_installing=self._artifact_installing,
+        )
+
+    def _enter_maintenance(self) -> None:
+        if self._module is not None:
+            self._module._clear_active = True
+
+    def _leave_maintenance(self) -> None:
+        if self._module is not None:
+            self._module._clear_active = False
+
+    def _resume_maintenance_claims(self) -> None:
+        if self._module is not None:
+            self._module._worker.resume_claims()
 
     def _configure_insight_reader(self, config: MemoryConfig) -> None:
         if self._insight_reader_override is not None:
@@ -438,8 +450,8 @@ class MemoryRuntime:
         try:
             return await self._reconcile(config)
         finally:
-            self._ensure_terminal_snapshot_gc()
-            self._ensure_backup_stage_reconcile()
+            if self._maintenance is not None:
+                self._maintenance.ensure_housekeeping()
 
     async def _reconcile(self, config: MemoryConfig) -> dict[str, Any]:
         """Run one reconciliation owned by the public lifecycle operation."""
@@ -474,17 +486,20 @@ class MemoryRuntime:
             # save cannot race a root wipe or replace sidecar credentials halfway
             # through an active provider call.
             async with self.module._lifecycle_lock:
-                restore_operation = self._open_backup_restore_operation()
+                restore_operation_open = (
+                    self._maintenance is not None
+                    and self._maintenance.has_open_restore()
+                )
                 maintenance_open = self._maintenance_open()
                 if (
                     maintenance_open
                     and self._can_disable_without_maintenance_authority(config)
                 ):
                     result = await self._disable_locked(config)
-                elif restore_operation is not None and not recorded_sidecar_reaped:
+                elif restore_operation_open and not recorded_sidecar_reaped:
                     self.module._worker.pause_claims()
                     return {"ok": False, "error": "memory_clear_failed"}
-                elif maintenance_open and restore_operation is None:
+                elif maintenance_open and not restore_operation_open:
                     self.module._worker.pause_claims()
                     return {"ok": False, "error": "memory_clear_failed"}
                 else:
@@ -523,8 +538,8 @@ class MemoryRuntime:
     ) -> dict[str, Any]:
         """Reconcile while both controller and module lifecycle locks are held."""
 
-        if self._open_backup_restore_operation() is not None:
-            recovered = await self._recover_backup_restore_locked()
+        if self._maintenance is not None and self._maintenance.has_open_restore():
+            recovered = await self._maintenance.recover_boot()
             if not recovered:
                 self.module._worker.pause_claims()
                 self._runtime_error = "memory_clear_failed"
@@ -845,32 +860,13 @@ class MemoryRuntime:
                     operator_ref=operator_ref
                 ),
             }
-        try:
-            meta, history, manual_required = await asyncio.gather(
-                asyncio.to_thread(self._store.get_meta),
-                asyncio.to_thread(self._store.has_provider_data_history),
-                asyncio.to_thread(self._store.has_manual_required_fence),
-            )
-        except Exception:
-            return {
-                "status": "ok",
-                "data_exists": True,
-                "can_clear": False,
-                "clear_recovery": self._clear_recovery_payload(
-                    operator_ref=operator_ref
-                ),
-            }
-        recovery = self._clear_recovery_payload(operator_ref=operator_ref)
+        maintenance = self._require_maintenance()
+        result = await maintenance.maintenance_payload(operator_ref=operator_ref)
         return {
             "status": "ok",
-            "data_exists": bool(history or (meta is not None and meta.last_success_at)),
-            "can_clear": (
-                self._clear_capability_ready()
-                and not self._maintenance_open()
-                and not manual_required
-                and recovery is None
-            ),
-            "clear_recovery": recovery,
+            "data_exists": result.data_exists,
+            "can_clear": result.can_clear,
+            "clear_recovery": _clear_recovery_payload(result.clear_recovery),
         }
 
     def principal_for_user_key(self, user_key: str) -> str:
@@ -1175,9 +1171,9 @@ class MemoryRuntime:
     async def create_backup(self, backup_id: str | None = None) -> MemorySnapshot:
         """Create one ordinary Memory backup under the full maintenance fence."""
 
-        return await self._run_backup_operation(
-            lambda manager: manager.create(backup_id),
-        )
+        if not self.available:
+            raise self._unavailable()
+        return await self._require_maintenance().create_backup(backup_id)
 
     async def restore_backup(
         self,
@@ -1190,346 +1186,72 @@ class MemoryRuntime:
 
         if not self.available:
             raise self._unavailable()
-        clear_journal = self._require_clear_journal()
-        restore_journal = self._require_backup_restore_journal()
-        manager = self._require_backup_manager()
-        async with self._reconcile_lock, self.module._lifecycle_lock:
-            async with self.module._root_lifecycle_lock():
-                clear_journal.assert_backup_allowed()
-                operation = restore_journal.get_open_operation()
-                if operation is not None:
-                    if (
-                        operation.state != "recovery_needed"
-                        or operation.backup_id != backup_id
-                        or operation.manifest_sha256 != expected_manifest_sha256
-                        or operation.digest_mapping() != dict(expected_surface_digests)
-                    ):
-                        raise BackupRestoreConflict(
-                            "a different Memory backup restore owns the recovery fence"
-                        )
-                self._backup_active = True
-                self.module._clear_active = True
-                try:
-                    await self._quiesce_for_clear()
-                    if operation is not None:
-                        operation = await self._run_maintenance_io(
-                            lambda: restore_journal.claim_retry(
-                                operation.operation_id,
-                                expected_revision=operation.revision,
-                                actor_ref="system:runtime",
-                            )
-                        )
-
-                    def begin_restore(snapshot: MemorySnapshot) -> None:
-                        nonlocal operation
-                        operation = restore_journal.start(snapshot)
-
-                    restored = await self._run_maintenance_io(
-                        lambda: manager.restore(
-                            backup_id,
-                            expected_manifest_sha256=expected_manifest_sha256,
-                            expected_surface_digests=expected_surface_digests,
-                            before_replace=None if operation is not None else begin_restore,
-                        )
-                    )
-                    if operation is None:
-                        raise RuntimeError("Memory backup restore intent was not recorded")
-                    operation = await self._run_maintenance_io(
-                        lambda: restore_journal.mark_completed(
-                            operation.operation_id,
-                            expected_revision=operation.revision,
-                            execution_token=self._backup_restore_execution_token(operation),
-                        )
-                    )
-                    return restored
-                except BaseException:
-                    await self._mark_backup_restore_recovery(operation)
-                    raise
-                finally:
-                    self._backup_active = False
-                    try:
-                        if restore_journal.get_open_operation() is None:
-                            await self._resume_after_clear()
-                    finally:
-                        self.module._clear_active = False
-
-    async def _run_backup_operation(
-        self,
-        operation: Callable[[MemorySnapshotManager], MemorySnapshot],
-    ) -> MemorySnapshot:
-        if not self.available:
-            raise self._unavailable()
-        journal = self._require_clear_journal()
-        restore_journal = self._require_backup_restore_journal()
-        manager = self._require_backup_manager()
-        async with self._reconcile_lock, self.module._lifecycle_lock:
-            async with self.module._root_lifecycle_lock():
-                journal.assert_backup_allowed()
-                restore_journal.assert_idle()
-                self._backup_active = True
-                self.module._clear_active = True
-                try:
-                    await self._quiesce_for_clear()
-                    return await self._run_maintenance_io(lambda: operation(manager))
-                finally:
-                    self._backup_active = False
-                    try:
-                        await self._resume_after_clear()
-                    finally:
-                        self.module._clear_active = False
-
-    @staticmethod
-    async def _run_maintenance_io(
-        operation: Callable[[], _MaintenanceIOResult],
-    ) -> _MaintenanceIOResult:
-        """Keep the maintenance fence until a cancelled I/O thread settles."""
-
-        return await run_blocking(operation)
+        return await self._require_maintenance().restore_backup(
+            backup_id,
+            expected_manifest_sha256=expected_manifest_sha256,
+            expected_surface_digests=expected_surface_digests,
+        )
 
     async def clear(self, *, operator_ref: str) -> dict[str, Any]:
         if not self.available:
             raise self._unavailable()
-        journal = self._require_clear_journal()
-        async with self._reconcile_lock, self.module._lifecycle_lock:
-            async with self.module._root_lifecycle_lock():
-                if journal.get_open_operation() is not None:
-                    return self._clear_blocked_payload(operator_ref=operator_ref)
-                operation: ClearOperation | None = None
-                self.module._clear_active = True
-                try:
-                    try:
-                        await self._pause_clear_claims()
-                        manual_required = await self._run_maintenance_io(
-                            self._store.has_manual_required_fence
-                        )
-                    except BaseException as error:
-                        self.module._clear_active = False
-                        self.module._worker.resume_claims()
-                        if isinstance(error, asyncio.CancelledError):
-                            raise
-                        return self._clear_blocked_payload(operator_ref=operator_ref)
-                    if manual_required:
-                        self.module._clear_active = False
-                        self.module._worker.resume_claims()
-                        return self._clear_blocked_payload(operator_ref=operator_ref)
-                    try:
-                        meta = await self._run_maintenance_io(self._store.ensure_meta)
-                        operation = await self._run_maintenance_io(
-                            lambda: journal.start(
-                                operator_ref=operator_ref,
-                                pre_epoch=meta.epoch,
-                                target_epoch=meta.epoch + 1,
-                            )
-                        )
-                        await self._run_maintenance_io(self._store.begin_clear_fence)
-                        operation = await self._prepare_clear(
-                            operation,
-                            claims_already_paused=True,
-                        )
-                        operation = await self._delete_clear_surfaces(operation)
-                    except BaseException as error:
-                        current = await self._mark_clear_recovery(operation)
-                        if current is not None and current.state == "completed":
-                            self.module._clear_active = False
-                            await self._finish_clear(current)
-                        elif current is None:
-                            self.module._clear_active = False
-                            self.module._worker.resume_claims()
-                        if isinstance(error, asyncio.CancelledError):
-                            raise
-                        return self._clear_blocked_payload(operator_ref=operator_ref)
-                finally:
-                    self.module._clear_active = False
-                if operation is None:
-                    raise RuntimeError("Memory clear operation did not start")
-                return await self._finish_clear(operation)
+        result = await self._require_maintenance().clear(operator_ref=operator_ref)
+        return _clear_result_payload(result)
 
-    async def resume_clear(self, operation_id: str, *, operator_ref: str) -> dict[str, Any]:
-        """Explicitly continue the exact clear stage recorded by the journal."""
-
-        if not self.available:
-            raise self._unavailable()
-        journal = self._require_clear_journal()
-        async with self._reconcile_lock, self.module._lifecycle_lock:
-            async with self.module._root_lifecycle_lock():
-                current = journal.get_open_operation()
-                if current is None or current.operation_id != operation_id:
-                    return self._clear_blocked_payload(operator_ref=operator_ref)
-                try:
-                    operation = await self._run_maintenance_io(
-                        lambda: journal.claim_resume(
-                            operation_id,
-                            operator_ref=operator_ref,
-                            expected_revision=current.revision,
-                        )
-                    )
-                    await self._run_maintenance_io(self._store.begin_clear_fence)
-                    operation = await self._prepare_clear(operation)
-                    operation = await self._delete_clear_surfaces(operation)
-                except BaseException as error:
-                    recovered = await self._mark_clear_recovery(
-                        operation_id=operation_id
-                    )
-                    if recovered is not None and recovered.state == "completed":
-                        await self._finish_clear(recovered)
-                    if isinstance(error, asyncio.CancelledError):
-                        raise
-                    return self._clear_blocked_payload(operator_ref=operator_ref)
-                return await self._finish_clear(operation)
-
-    async def abort_clear(self, operation_id: str, *, operator_ref: str) -> dict[str, Any]:
-        """Restore all four verified surfaces, then release the admission fence."""
-
-        if not self.available:
-            raise self._unavailable()
-        journal = self._require_clear_journal()
-        manager = self._require_snapshot_manager()
-        async with self._reconcile_lock, self.module._lifecycle_lock:
-            async with self.module._root_lifecycle_lock():
-                current = journal.get_open_operation()
-                if current is None or current.operation_id != operation_id:
-                    return self._clear_blocked_payload(operator_ref=operator_ref)
-                try:
-                    operation = await self._run_maintenance_io(
-                        lambda: journal.claim_abort(
-                            operation_id,
-                            operator_ref=operator_ref,
-                            expected_revision=current.revision,
-                        )
-                    )
-                    await self._quiesce_for_clear()
-                    digests = self._journal_surface_digests(operation.operation_id)
-                    if operation.manifest_sha256 is None or operation.snapshot_path is None:
-                        raise RuntimeError("clear snapshot is not sealed")
-                    await self._run_maintenance_io(
-                        lambda: manager.restore(
-                            operation.operation_id,
-                            expected_manifest_sha256=operation.manifest_sha256,
-                            expected_surface_digests=digests,
-                        )
-                    )
-                    await self._run_maintenance_io(self._store.release_clear_fence)
-                    for surface in journal.get_surfaces(operation.operation_id):
-                        if surface.state == "restored":
-                            continue
-                        operation = await self._run_maintenance_io(
-                            lambda: journal.record_surface_restored(
-                                operation.operation_id,
-                                surface.surface,
-                                expected_revision=operation.revision,
-                                execution_token=self._clear_execution_token(operation),
-                            )
-                        )
-                    operation = await self._run_maintenance_io(
-                        lambda: journal.mark_aborted(
-                            operation.operation_id,
-                            expected_revision=operation.revision,
-                            execution_token=self._clear_execution_token(operation),
-                        )
-                    )
-                except BaseException as error:
-                    recovered = await self._mark_clear_recovery(
-                        operation_id=operation_id
-                    )
-                    if recovered is not None and recovered.state == "aborted":
-                        await self._finish_aborted_clear(recovered)
-                    if isinstance(error, asyncio.CancelledError):
-                        raise
-                    return self._clear_blocked_payload(operator_ref=operator_ref)
-                return await self._finish_aborted_clear(operation)
-
-    async def _prepare_clear(
+    async def resume_clear(
         self,
-        operation: ClearOperation,
+        operation_id: str,
         *,
-        claims_already_paused: bool = False,
-    ) -> ClearOperation:
-        journal = self._require_clear_journal()
-        manager = self._require_snapshot_manager()
-        await self._quiesce_for_clear(
-            claims_already_paused=claims_already_paused,
+        operator_ref: str,
+    ) -> dict[str, Any]:
+        if not self.available:
+            raise self._unavailable()
+        result = await self._require_maintenance().resume_clear(
+            operation_id,
+            operator_ref=operator_ref,
         )
-        if operation.state == "preparing":
-            if operation.snapshot_path is None or operation.manifest_sha256 is None:
-                if operation.resolution == "resume":
-                    with journal.authorize_preparing_snapshot_discard(
-                        operation.operation_id,
-                        expected_revision=operation.revision,
-                        execution_token=self._clear_execution_token(operation),
-                    ) as permit:
-                        await self._run_maintenance_io(
-                            lambda: manager.discard_unrecorded(permit)
-                        )
-                    refreshed = journal.get_operation(operation.operation_id)
-                    if refreshed is None:
-                        raise RuntimeError("Memory clear operation disappeared")
-                    operation = refreshed
-                snapshot = await self._run_maintenance_io(
-                    lambda: manager.create(operation.operation_id)
-                )
-                operation = await self._run_maintenance_io(
-                    lambda: journal.record_snapshot(
-                        operation.operation_id,
-                        expected_revision=operation.revision,
-                        execution_token=self._clear_execution_token(operation),
-                        snapshot=snapshot,
-                    )
-                )
-            else:
-                await self._verify_clear_snapshot(operation)
-            operation = await self._run_maintenance_io(
-                lambda: journal.mark_prepared(
-                    operation.operation_id,
-                    expected_revision=operation.revision,
-                    execution_token=self._clear_execution_token(operation),
-                )
-            )
-        if operation.state == "prepared":
-            await self._verify_clear_snapshot(operation)
-        return operation
+        return _clear_result_payload(result)
 
-    async def _delete_clear_surfaces(self, operation: ClearOperation) -> ClearOperation:
-        journal = self._require_clear_journal()
-        if operation.state == "prepared":
-            operation = await self._run_maintenance_io(
-                lambda: journal.begin_deleting(
-                    operation.operation_id,
-                    expected_revision=operation.revision,
-                    execution_token=self._clear_execution_token(operation),
-                )
-            )
-        if operation.state != "deleting":
-            raise RuntimeError("clear operation is not ready for deletion")
-        await self._verify_clear_snapshot(operation)
-        for surface in journal.get_surfaces(operation.operation_id):
-            if surface.state == "deleted":
-                continue
-            await self._delete_clear_surface(surface, target_epoch=operation.target_epoch)
-            operation = await self._run_maintenance_io(
-                lambda: journal.record_surface_deleted(
-                    operation.operation_id,
-                    surface.surface,
-                    expected_revision=operation.revision,
-                    execution_token=self._clear_execution_token(operation),
-                )
-            )
-        return await self._run_maintenance_io(
-            lambda: journal.mark_completed(
-                operation.operation_id,
-                expected_revision=operation.revision,
-                execution_token=self._clear_execution_token(operation),
-            )
+    async def abort_clear(
+        self,
+        operation_id: str,
+        *,
+        operator_ref: str,
+    ) -> dict[str, Any]:
+        if not self.available:
+            raise self._unavailable()
+        result = await self._require_maintenance().abort_clear(
+            operation_id,
+            operator_ref=operator_ref,
         )
+        return _clear_result_payload(result)
+
+    async def _pause_clear_claims(self) -> None:
+        if not await self.module._worker.pause_and_wait(
+            timeout_seconds=self.module._clear_drain_timeout_seconds
+        ):
+            raise RuntimeError("Memory worker did not quiesce before clear")
+
+    async def _quiesce_for_clear(self, claims_already_paused: bool = False) -> None:
+        if not claims_already_paused:
+            await self._pause_clear_claims()
+        await self._stop_worker()
+        if self._process is not None:
+            await self._process.stop()
+            self._process = None
+        self._process_records_calls = False
+        await self._stop_call_log_retention()
+        self._set_recorder_health_disabled()
 
     async def _delete_clear_surface(
         self,
         surface: ClearSurface,
-        *,
         target_epoch: int,
     ) -> None:
         if surface.surface == "queue":
-            await self._run_maintenance_io(
-                lambda: self._store.reset_for_clear(target_epoch=target_epoch)
+            await run_blocking(
+                self._store.reset_for_clear,
+                target_epoch=target_epoch,
             )
             return
         if surface.surface == "provider":
@@ -1547,365 +1269,16 @@ class MemoryRuntime:
                         self._active_provider_root_metadata(),
                     )
 
-            await self._run_maintenance_io(reset_provider_root)
+            await run_blocking(reset_provider_root)
             return
         if surface.surface == "call_log":
-            await self._run_maintenance_io(
-                lambda: clear_call_log(self._call_log_db_path)
-            )
+            await run_blocking(clear_call_log, self._call_log_db_path)
             self._set_recorder_health_disabled()
             return
         if surface.surface == "attachments":
-            await self._run_maintenance_io(
-                self.module._attachment_store.clear_all
-            )
+            await run_blocking(self.module._attachment_store.clear_all)
             return
         raise RuntimeError("unknown Memory clear surface")
-
-    async def _verify_clear_snapshot(self, operation: ClearOperation) -> MemorySnapshot:
-        if operation.manifest_sha256 is None or operation.snapshot_path is None:
-            raise RuntimeError("clear snapshot is not sealed")
-        return await self._run_maintenance_io(
-            lambda: self._require_snapshot_manager().verify(
-                operation.operation_id,
-                expected_manifest_sha256=operation.manifest_sha256,
-                expected_surface_digests=self._journal_surface_digests(
-                    operation.operation_id
-                ),
-            )
-        )
-
-    async def _pause_clear_claims(self) -> None:
-        if not await self.module._worker.pause_and_wait(
-            timeout_seconds=self.module._clear_drain_timeout_seconds
-        ):
-            raise RuntimeError("Memory worker did not quiesce before clear")
-
-    async def _quiesce_for_clear(self, *, claims_already_paused: bool = False) -> None:
-        if not claims_already_paused:
-            await self._pause_clear_claims()
-        await self._stop_worker()
-        if self._process is not None:
-            await self._process.stop()
-            self._process = None
-        self._process_records_calls = False
-        await self._stop_call_log_retention()
-        self._set_recorder_health_disabled()
-
-    async def _recover_backup_restore_locked(self) -> bool:
-        """Converge an interrupted restore before any boot-time activation."""
-
-        journal = self._require_backup_restore_journal()
-        operation = journal.get_open_operation()
-        if operation is None:
-            return True
-        if operation.state != "recovery_needed":
-            return False
-        manager = self._require_backup_manager()
-        self._backup_active = True
-        self.module._clear_active = True
-        try:
-            async with self.module._root_lifecycle_lock():
-                await self._quiesce_for_clear()
-                operation = await self._run_maintenance_io(
-                    lambda: journal.claim_retry(
-                        operation.operation_id,
-                        expected_revision=operation.revision,
-                        actor_ref="system:boot",
-                    )
-                )
-                await self._run_maintenance_io(
-                    lambda: manager.restore(
-                        operation.backup_id,
-                        expected_manifest_sha256=operation.manifest_sha256,
-                        expected_surface_digests=operation.digest_mapping(),
-                    )
-                )
-                await self._run_maintenance_io(
-                    lambda: journal.mark_completed(
-                        operation.operation_id,
-                        expected_revision=operation.revision,
-                        execution_token=self._backup_restore_execution_token(operation),
-                        actor_ref="system:boot",
-                    )
-                )
-            return True
-        except asyncio.CancelledError:
-            await self._mark_backup_restore_recovery(operation)
-            raise
-        except Exception:
-            await self._mark_backup_restore_recovery(operation)
-            return False
-        finally:
-            self._backup_active = False
-            self.module._clear_active = False
-
-    async def _mark_backup_restore_recovery(
-        self,
-        operation: BackupRestoreOperation | None,
-    ) -> BackupRestoreOperation | None:
-        journal = self._backup_restore_journal
-        if journal is None or operation is None:
-            return None
-        try:
-            current = journal.get_operation(operation.operation_id)
-            if current is None or current.state != "restoring":
-                return current
-            return await self._run_maintenance_io(
-                lambda: journal.mark_recovery_needed(
-                    current.operation_id,
-                    expected_revision=current.revision,
-                    execution_token=self._backup_restore_execution_token(current),
-                )
-            )
-        except Exception:
-            logger.exception("Memory backup restore failure could not be journaled")
-            return None
-
-    async def _mark_clear_recovery(
-        self,
-        operation: ClearOperation | None = None,
-        *,
-        operation_id: str | None = None,
-    ) -> ClearOperation | None:
-        journal = self._clear_journal
-        if journal is None:
-            return None
-
-        def mark_recovery() -> ClearOperation | None:
-            identifier = operation.operation_id if operation is not None else operation_id
-            current = (
-                journal.get_open_operation()
-                if identifier is None
-                else journal.get_operation(identifier)
-            )
-            if current is None:
-                return None
-            if current.state in {"preparing", "prepared", "deleting"}:
-                return journal.mark_recovery_needed(
-                    current.operation_id,
-                    expected_revision=current.revision,
-                    execution_token=self._clear_execution_token(current),
-                )
-            if current.state == "recovery_needed" and current.execution_token is not None:
-                return journal.release_recovery_claim(
-                    current.operation_id,
-                    expected_revision=current.revision,
-                    execution_token=self._clear_execution_token(current),
-                )
-            return current
-
-        try:
-            return await self._run_maintenance_io(mark_recovery)
-        except Exception:
-            logger.exception("Memory clear failure could not be journaled")
-            return None
-
-    async def _finish_clear(self, operation: ClearOperation) -> dict[str, Any]:
-        try:
-            await self._run_maintenance_io(self._reconcile_terminal_clear_snapshots)
-        finally:
-            await self._resume_after_clear()
-        return {
-            "status": "completed",
-            "operation_id": operation.operation_id,
-            "epoch": operation.target_epoch,
-        }
-
-    async def _finish_aborted_clear(self, operation: ClearOperation) -> dict[str, Any]:
-        try:
-            await self._run_maintenance_io(self._reconcile_terminal_clear_snapshots)
-        finally:
-            await self._resume_after_clear()
-        return {
-            "status": "aborted",
-            "operation_id": operation.operation_id,
-            "epoch": operation.pre_epoch,
-        }
-
-    def _reconcile_terminal_clear_snapshots(self) -> None:
-        """Retry GC whose durable terminal transition preceded snapshot removal."""
-
-        journal = self._require_clear_journal()
-        manager = self._require_snapshot_manager()
-        for permit in journal.terminal_snapshot_permits():
-            try:
-                manager.remove(permit)
-            except Exception:
-                logger.warning(
-                    "Terminal Memory clear snapshot %s could not be removed",
-                    permit.snapshot_id,
-                    exc_info=True,
-                )
-
-    def _ensure_terminal_snapshot_gc(self) -> None:
-        """Schedule one runtime-owned retry after activation has returned."""
-
-        task = self._terminal_snapshot_gc_task
-        if (
-            self._closing
-            or self._artifact_installing
-            or self._clear_journal is None
-            or self._snapshot_manager is None
-            or self._clear_journal_error is not None
-            or (task is not None and not task.done())
-        ):
-            return
-        task = asyncio.create_task(
-            self._run_terminal_snapshot_gc(),
-            name="memory-terminal-snapshot-gc",
-        )
-        self._terminal_snapshot_gc_task = task
-
-        def clear_gc(completed: asyncio.Task[None]) -> None:
-            if self._terminal_snapshot_gc_task is completed:
-                self._terminal_snapshot_gc_task = None
-
-        task.add_done_callback(clear_gc)
-
-    def _ensure_backup_stage_reconcile(self) -> None:
-        """Schedule cleanup of auto-ID backup stages after lifecycle reconcile."""
-
-        task = self._backup_stage_reconcile_task
-        if (
-            self._closing
-            or self._artifact_installing
-            or not self.available
-            or self._backup_manager is None
-            or self._clear_journal is None
-            or self._backup_restore_journal is None
-            or self._clear_journal_error is not None
-            or (task is not None and not task.done())
-        ):
-            return
-        task = asyncio.create_task(
-            self._run_backup_stage_reconcile(),
-            name="memory-backup-stage-reconcile",
-        )
-        self._backup_stage_reconcile_task = task
-
-        def clear_reconcile(completed: asyncio.Task[None]) -> None:
-            if self._backup_stage_reconcile_task is completed:
-                self._backup_stage_reconcile_task = None
-
-        task.add_done_callback(clear_reconcile)
-
-    async def _run_backup_stage_reconcile(self) -> None:
-        """Remove abandoned unpublished stages under lifecycle serialization."""
-
-        # Both jobs reopen the clear journal, whose transient SQLite sidecars
-        # must not be inspected while the preceding terminal GC is settling.
-        terminal_gc = self._terminal_snapshot_gc_task
-        if terminal_gc is not None and terminal_gc is not asyncio.current_task():
-            await asyncio.shield(terminal_gc)
-        journal = self._require_clear_journal()
-        restore_journal = self._require_backup_restore_journal()
-        manager = self._require_backup_manager()
-        try:
-            async with self._reconcile_lock, self.module._lifecycle_lock:
-                async with self.module._root_lifecycle_lock():
-                    if self._artifact_installing:
-                        return
-                    journal.assert_backup_allowed()
-                    restore_journal.assert_idle()
-                    await self._run_maintenance_io(
-                        manager.reconcile_unpublished_backup_stages
-                    )
-        except asyncio.CancelledError:
-            raise
-        except Exception:
-            logger.warning(
-                "Unpublished Memory backup stages could not be reconciled",
-                exc_info=True,
-            )
-
-    async def _stop_backup_stage_reconcile(self) -> None:
-        """Cancel stage cleanup and wait until its current filesystem I/O settles."""
-
-        task = self._backup_stage_reconcile_task
-        if task is None or task is asyncio.current_task():
-            return
-        task.cancel()
-        settlement = asyncio.gather(task, return_exceptions=True)
-        caller_cancellation: asyncio.CancelledError | None = None
-        while not settlement.done():
-            try:
-                await asyncio.shield(settlement)
-            except asyncio.CancelledError as error:
-                caller_cancellation = caller_cancellation or error
-        try:
-            result = settlement.result()[0]
-        finally:
-            if self._backup_stage_reconcile_task is task:
-                self._backup_stage_reconcile_task = None
-        if isinstance(result, BaseException) and not isinstance(
-            result,
-            asyncio.CancelledError,
-        ):
-            raise result
-        if caller_cancellation is not None:
-            raise caller_cancellation
-
-    async def _run_terminal_snapshot_gc(self) -> None:
-        """Remove terminal snapshots without occupying the activation path."""
-
-        journal = self._require_clear_journal()
-        manager = self._require_snapshot_manager()
-        try:
-            permits = await self._run_maintenance_io(
-                journal.terminal_snapshot_permits
-            )
-        except asyncio.CancelledError:
-            raise
-        except Exception:
-            logger.warning(
-                "Terminal Memory clear snapshot permits could not be loaded",
-                exc_info=True,
-            )
-            return
-        for permit in permits:
-            try:
-                # A cancellation settles the current filesystem operation but
-                # stops before the next permit. Remaining permits stay durable
-                # for the next reconcile or restart.
-                await self._run_maintenance_io(
-                    lambda permit=permit: manager.remove(permit)
-                )
-            except asyncio.CancelledError:
-                raise
-            except Exception:
-                logger.warning(
-                    "Terminal Memory clear snapshot %s could not be removed",
-                    permit.snapshot_id,
-                    exc_info=True,
-                )
-
-    async def _stop_terminal_snapshot_gc(self) -> None:
-        """Cancel GC and retain ownership until its current I/O call settles."""
-
-        task = self._terminal_snapshot_gc_task
-        if task is None or task is asyncio.current_task():
-            return
-        task.cancel()
-        settlement = asyncio.gather(task, return_exceptions=True)
-        caller_cancellation: asyncio.CancelledError | None = None
-        while not settlement.done():
-            try:
-                await asyncio.shield(settlement)
-            except asyncio.CancelledError as error:
-                caller_cancellation = caller_cancellation or error
-        try:
-            result = settlement.result()[0]
-        finally:
-            if self._terminal_snapshot_gc_task is task:
-                self._terminal_snapshot_gc_task = None
-        if isinstance(result, BaseException) and not isinstance(
-            result,
-            asyncio.CancelledError,
-        ):
-            raise result
-        if caller_cancellation is not None:
-            raise caller_cancellation
 
     async def _resume_after_clear(self) -> None:
         """Keep lifecycle ownership until runtime reconciliation settles."""
@@ -1943,107 +1316,24 @@ class MemoryRuntime:
         else:
             self._runtime_error = "memory_sidecar_unavailable"
 
-    def _open_backup_restore_operation(self) -> BackupRestoreOperation | None:
-        journal = self._backup_restore_journal
-        if journal is None:
-            return None
-        try:
-            return journal.get_open_operation()
-        except Exception:
-            return None
-
-    def _open_clear_operation(self) -> ClearOperation | None:
-        journal = self._clear_journal
-        if journal is None:
-            return None
-        try:
-            return journal.get_open_operation()
-        except Exception:
-            return None
-
     def _clear_recovery_payload(
         self,
         *,
         operator_ref: str | None = None,
     ) -> dict[str, Any] | None:
-        operation = self._open_clear_operation()
-        if operation is None:
-            return None
-        can_resume = False
-        can_abort = False
-        try:
-            if operator_ref is not None and operation.operator_ref == operator_ref:
-                journal = self._require_clear_journal()
-                can_resume = journal.can_resume(operation.operation_id)
-                can_abort = journal.can_abort(operation.operation_id)
-        except Exception:
-            pass
-        return {
-            "state": operation.state,
-            "operation_id": operation.operation_id,
-            "occurred_at": operation.updated_at,
-            "error_code": operation.closed_error,
-            "can_resume": can_resume,
-            "can_abort": can_abort,
-        }
+        maintenance = self._maintenance
+        recovery = (
+            None
+            if maintenance is None
+            else maintenance.recovery(operator_ref=operator_ref)
+        )
+        return _clear_recovery_payload(recovery)
 
-    def _clear_blocked_payload(
-        self,
-        *,
-        operator_ref: str | None = None,
-    ) -> dict[str, Any]:
-        return {
-            "status": "failed",
-            "error": "memory_clear_failed",
-            "recovery": self._clear_recovery_payload(operator_ref=operator_ref),
-        }
-
-    def _journal_surface_digests(self, operation_id: str) -> dict[str, str | None]:
-        journal = self._require_clear_journal()
-        return {
-            surface.relative_path: surface.snapshot_digest
-            for surface in journal.get_surfaces(operation_id)
-        }
-
-    def _require_clear_journal(self) -> MemoryClearJournal:
-        if self._clear_journal is None or self._clear_journal_error is not None:
-            raise MemoryStoreUnavailableError("Memory clear journal is unavailable")
-        return self._clear_journal
-
-    def _require_backup_restore_journal(self) -> MemoryBackupRestoreJournal:
-        if (
-            self._backup_restore_journal is None
-            or self._clear_journal_error is not None
-        ):
-            raise MemoryStoreUnavailableError(
-                "Memory backup restore journal is unavailable"
-            )
-        return self._backup_restore_journal
-
-    def _require_snapshot_manager(self) -> MemorySnapshotManager:
-        if self._snapshot_manager is None or self._clear_journal_error is not None:
-            raise MemoryStoreUnavailableError("Memory snapshot manager is unavailable")
-        return self._snapshot_manager
-
-    def _require_backup_manager(self) -> MemorySnapshotManager:
-        if self._backup_manager is None or self._clear_journal_error is not None:
-            raise MemoryStoreUnavailableError("Memory backup manager is unavailable")
-        return self._backup_manager
-
-    @staticmethod
-    def _backup_restore_execution_token(
-        operation: BackupRestoreOperation,
-    ) -> str:
-        if operation.execution_token is None:
-            raise RuntimeError("Memory backup restore execution claim is missing")
-        return operation.execution_token
-
-    @staticmethod
-    def _clear_execution_token(operation: ClearOperation) -> str:
-        if operation.execution_token is None:
-            raise RuntimeError("Memory clear operation is not claimed")
-        return operation.execution_token
-
+    def _require_maintenance(self) -> MemoryMaintenance:
+        maintenance = self._maintenance
+        if maintenance is None:
+            raise self._unavailable()
+        return maintenance
     async def install_artifact(self) -> dict[str, Any]:
         """Install or repair EverOS through this controller-owned lifecycle."""
 
@@ -2137,8 +1427,8 @@ class MemoryRuntime:
                         self._artifact_installing = False
                 except asyncio.CancelledError as error:
                     cancellation = cancellation or error
-            self._ensure_terminal_snapshot_gc()
-            self._ensure_backup_stage_reconcile()
+            if self._maintenance is not None:
+                self._maintenance.ensure_housekeeping()
         if cancellation is not None:
             raise cancellation
         if ensure_failed:
@@ -2184,8 +1474,8 @@ class MemoryRuntime:
             logger.exception("Memory sidecar restart failed")
             return {"ok": False, "error": "memory_restart_failed"}
         finally:
-            self._ensure_terminal_snapshot_gc()
-            self._ensure_backup_stage_reconcile()
+            if self._maintenance is not None:
+                self._maintenance.ensure_housekeeping()
 
     async def _restart_locked(self) -> dict[str, Any]:
         """Replace the sidecar while ``_reconcile_lock`` is held."""
@@ -2320,7 +1610,8 @@ class MemoryRuntime:
 
     async def close(self) -> None:
         self._closing = True
-        await self._stop_backup_stage_reconcile()
+        if self._maintenance is not None:
+            await self._maintenance.close()
         restart_task = self._restart_task
         if restart_task is not None and restart_task is not asyncio.current_task():
             try:
@@ -2353,7 +1644,6 @@ class MemoryRuntime:
         self._process_records_calls = False
         await self._stop_call_log_retention()
         self._artifact_manager.set_activation_coordinator(None)
-        await self._stop_terminal_snapshot_gc()
 
     def _active_provider_root_metadata(self) -> ProviderRootMetadata:
         provider_root_format = (
